refactor: update import paths for my-filters components
- Refactored import statements for various components within the my-filters feature, moving them from the `my-filters` directory to a new `features/my-filters` directory.
- This change enhances code organization and maintainability by providing a clearer structure for the components related to filters.
