+++ src/app/(app)/(stacked-layout)/my-filters/new-filter/loading.tsx
@@ -1,4 +1,4 @@
-import { FormSkeleton } from "@/components/my-filters/form-skeleton";
+import { FormSkeleton } from "@/components/features/my-filters/components/form-skeleton";
 
 export default function Loading() {
   return <FormSkeleton />;
