+++ src/app/(app)/(stacked-layout)/my-filters/new-filter/new-filter-form.tsx
@@ -25,8 +25,8 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { ConveyorCard } from "@/components/conveyor-card";
+import { FilterCategoryCombobox } from "@/components/features/my-filters/components/filter-category-combobox";
 import { FilterImageCombobox } from "@/components/filter-image-combobox";
-import { FilterCategoryCombobox } from "@/components/my-filters/filter-category-combobox";
 
 const DevTool = dynamic(
   () => import("@hookform/devtools").then((module) => module.DevTool),
