+++ src/components/features/my-filters/new-filter/clear-input-button.tsx
