+++ src/components/features/my-filters/shared-filters/private-share-dropdown-item.tsx
@@ -1,6 +1,6 @@
 import { Share2Icon } from "lucide-react";
 
-import { DropdownMenuItem } from "../../ui/dropdown-menu";
+import { DropdownMenuItem } from "@/components/ui/dropdown-menu";
 
 interface PrivateShareDropdownItemProps {
   filterId: number;
